modify downstream code

+++ bindings/java/rust_code/src/lib.rs
@@ -14,10 +14,7 @@ pub extern "system" fn Java_ethereum_cryptography_LibEthKZG_DASContextNew(
     num_threads: jlong,
 ) -> jlong {
     let use_precomp = use_precomp != 0;
-    // On a 16bit system, the `as usize` will truncate to a u16
-    // This is fine because we do not expect a 16 bit system to be
-    // passing in more than 2^16 number of threads as an argument.
-    let num_threads = (num_threads as u64) as usize;
+    let num_threads = (num_threads as u64) as u32;
     c_eth_kzg::eth_kzg_das_context_new(use_precomp, num_threads) as jlong
 }
